gpt-main/gpt-main/backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Data processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
xlrd==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
bcrypt==4.1.2

# Utilities
psutil==5.9.8
python-dateutil==2.8.2
pytz==2024.1
python-json-logger==2.0.7
email-validator==2.1.0
requests==2.31.0

# HTTP client
httpx==0.26.0
aiofiles==23.2.1

# Rate limiting
slowapi==0.1.9
